Refactor: extract Calibre activity logging from ModelData to CalibreActivityLogger
- Created CalibreActivityLogger to handle background Realm logging of Calibre activities.
- Migrated activityDispatchQueue, logCalibreActivitySubject, and related logic to the new logger.
- Injected CalibreActivityLogger into ModelData and initialized it with realmConf.
- Updated CalibreServerService to use the new logger instance.
- Cleaned up redundant logging code from ModelData while maintaining backward compatibility.
