+++ YetAnotherEBookReader/Models/ModelData.swift
@@ -124,7 +124,7 @@ final class ModelData: ObservableObject {
     var realmSaveBooksMetadata: Realm!
     var realmConf: Realm.Configuration!
     
-    let activityDispatchQueue = DispatchQueue(label: "io.github.dsreader.activity")
+    var logger: CalibreActivityLogger!
     
     let kfImageCache = ImageCache.default
     var authResponsor = AuthResponsor()
@@ -168,8 +168,6 @@ final class ModelData: ObservableObject {
     let calibreUpdatedSubject = PassthroughSubject<calibreUpdatedSignal, Never>()
     let setLastReadPositionSubject = PassthroughSubject<CalibreBookSetLastReadPositionTask, Never>()
     let updateAnnotationsSubject = PassthroughSubject<CalibreBookUpdateAnnotationsTask, Never>()
-    
-    let logCalibreActivitySubject = PassthroughSubject<CalibreActivity, Never>()
     
     @Published var librarySyncStatus = [String: CalibreSyncStatus]()
 
@@ -221,8 +219,6 @@ final class ModelData: ObservableObject {
         registerBookReaderClosedCancellable()
         
         registerRecentShelfUpdater()
-        
-        registerLogCalibreActivityCancellable()
         
         bookDownloadedSubject.sink { book in
             self.calibreUpdatedSubject.send(.book(book))
@@ -497,6 +493,7 @@ final class ModelData: ObservableObject {
         realm = try! Realm(
             configuration: realmConf
         )
+        logger = CalibreActivityLogger(realmConf: realmConf)
         ModelData.SaveBooksMetadataRealmQueue.sync {
             self.realmSaveBooksMetadata = try! Realm(
                 configuration: realmConf, queue: ModelData.SaveBooksMetadataRealmQueue
@@ -2514,100 +2511,11 @@ final class ModelData: ObservableObject {
     }
     
     func logStartCalibreActivity(type: String, request: URLRequest, startDatetime: Date, bookId: Int32?, libraryId: String?) {
-        logCalibreActivitySubject.send(
-            CalibreActivityStart(type, request, startDatetime: startDatetime, bookId: bookId, libraryId: libraryId)
-        )
-//        activityDispatchQueue.async {
-//            guard let realm = try? Realm(configuration: self.realmConf) else { return }
-//
-//            let obj = CalibreActivityLogEntry()
-//
-//            obj.type = type
-//
-//            obj.startDatetime = startDatetime
-//            obj.bookId = bookId ?? 0
-//            obj.libraryId = libraryId
-//
-//            obj.endpoingURL = request.url?.absoluteString
-//            obj.httpMethod = request.httpMethod
-//            obj.httpBody = request.httpBody
-//            request.allHTTPHeaderFields?.forEach {
-//                obj.requestHeaders.append($0.key)
-//                obj.requestHeaders.append($0.value)
-//            }
-//
-//            try? realm.write {
-//                realm.add(obj)
-//            }
-//        }
+        logger.logStartCalibreActivity(type: type, request: request, startDatetime: startDatetime, bookId: bookId, libraryId: libraryId)
     }
     
     func logFinishCalibreActivity(type: String, request: URLRequest, startDatetime: Date, finishDatetime: Date, errMsg: String) {
-        logCalibreActivitySubject.send(
-            CalibreActivityFinish(type, request, startDatetime: startDatetime, finishDatetime: finishDatetime, errMsg: errMsg)
-        )
-//        activityDispatchQueue.async {
-//            guard let realm = try? Realm(configuration: self.realmConf) else { return }
-//
-//            guard let activity = realm.objects(CalibreActivityLogEntry.self).filter(
-//                NSPredicate(format: "type = %@ AND startDatetime = %@ AND endpoingURL = %@",
-//                            type,
-//                            startDatetime as NSDate,
-//                            request.url?.absoluteString ?? ""
-//                )
-//            ).first else { return }
-//
-//            try? realm.write {
-//                activity.finishDatetime = finishDatetime
-//                activity.errMsg = errMsg
-//            }
-//        }
-    }
-    
-    func registerLogCalibreActivityCancellable() {
-        logCalibreActivitySubject.collect(.byTimeOrCount(RunLoop.main, .seconds(1), 16))
-            .receive(on: activityDispatchQueue)
-            .sink { activities in
-                guard let realm = try? Realm(configuration: self.realmConf) else { return }
-                
-                try? realm.write {
-                    activities.forEach { activity in
-                        if let activityStart = activity as? CalibreActivityStart {
-                            let obj = CalibreActivityLogEntry()
-                            
-                            obj.type = activityStart.type
-                            
-                            obj.startDatetime = activityStart.startDatetime
-                            obj.bookId = activityStart.bookId ?? 0
-                            obj.libraryId = activityStart.libraryId
-                            
-                            obj.endpoingURL = activityStart.request.url?.absoluteString
-                            obj.httpMethod = activityStart.request.httpMethod
-                            obj.httpBody = activityStart.request.httpBody
-                            activityStart.request.allHTTPHeaderFields?.forEach {
-                                obj.requestHeaders.append($0.key)
-                                obj.requestHeaders.append($0.value)
-                            }
-                            
-                            realm.add(obj)
-                        }
-                        
-                        if let activityFinish = activity as? CalibreActivityFinish {
-                            guard let activityPrevious = realm.objects(CalibreActivityLogEntry.self).filter(
-                                NSPredicate(format: "type = %@ AND startDatetime = %@ AND endpoingURL = %@",
-                                            activityFinish.type,
-                                            activityFinish.startDatetime as NSDate,
-                                            activityFinish.request.url?.absoluteString ?? ""
-                                )
-                            ).first else { return }
-                            
-                            activityPrevious.finishDatetime = activityFinish.finishDatetime
-                            activityPrevious.errMsg = activityFinish.errMsg
-                        }
-                    }
-                }
-            }
-            .store(in: &calibreCancellables)
+        logger.logFinishCalibreActivity(type: type, request: request, startDatetime: startDatetime, finishDatetime: finishDatetime, errMsg: errMsg)
     }
     
     func removeCalibreActivity(obj: CalibreActivityLogEntry) {
